Moved paste module into one file
